Fix build (#197)
See Changelog for details.
