+++ .gitignore
@@ -105,3 +105,4 @@ venv.bak/
 
 .vscode/
 .idea
+panoptes/utils/version.py
