--- MANIFEST.in
@@ -1,3 +1,0 @@
-include versioneer.py
-include panoptes/utils/_version.py
-include panoptes/utils/stars.py
