+++ conftest.py
@@ -18,7 +18,7 @@ import logging
 from _pytest.logging import caplog as _caplog
 from contextlib import suppress
 
-from panoptes.utils.logger import logger
+from panoptes.utils.logging import logger
 from panoptes.utils.database import PanDB
 from panoptes.utils.messaging import PanMessaging
 from panoptes.utils.config.client import set_config
