+++ panoptes/utils/__init__.py
@@ -1,9 +1,5 @@
-from .logger import logger
+from .logging import logger
 from .utils import *
 from .time import *
-
-from ._version import get_versions
-__version__ = get_versions()['version']
-del get_versions
 
 logger.disable('panoptes')
