+++ panoptes/utils/config/client.py
@@ -1,6 +1,6 @@
 import requests
 
-from ..logger import logger
+from ..logging import logger
 from ..serializers import from_json
 from ..serializers import to_json
