+++ panoptes/utils/tests/serial_handlers/protocol_arduinosimulator.py
@@ -14,7 +14,7 @@ import threading
 import time
 import urllib
 
-from panoptes.utils.logger import logger
+from panoptes.utils.logging import logger
 from panoptes.utils.tests import serial_handlers
 from panoptes.utils.serializers import to_json
 from panoptes.utils.serializers import from_json
