+++ panoptes/utils/theskyx.py
@@ -1,6 +1,6 @@
 import socket
 
-from .logger import logger
+from .logging import logger
 from . import error
