+++ setup.cfg
@@ -1,5 +1,18 @@
+[metadata]
+author = PANOPTES Team
+author_email = info@projectpanoptes.org
+description = Astronomical utilities. PANOPTES Style.
+edit_on_github = True
+github_project = panoptes/panoptes-utils
+keywords = exoplanet digital dslr camera utilities astronomy
+license = MIT
+long_description = A set of utility modules, functions, and services. Designed for PANOPTES, maybe useful for others.
+name = panoptes-utils
+package_name = panoptes.utils
+url = http://github.com/panoptes/panoptes-utils
+
 [aliases]
-test=pytest
+test = pytest
 
 [build_sphinx]
 source-dir = docs
@@ -14,7 +27,7 @@ packages =
     panoptes-utils
 
 [tool:pytest]
-python_files= test_*.py
+python_files = test_*.py
 norecursedirs = scripts resources bin docker
 addopts = --doctest-modules
 doctest_optionflags = ELLIPSIS NORMALIZE_WHITESPACE ALLOW_UNICODE IGNORE_EXCEPTION_DETAIL
@@ -23,26 +36,9 @@ filterwarnings =
     ignore::pytest.PytestDeprecationWarning
 doctest_plus = enabled
 
-[metadata]
-author = PANOPTES Team
-author_email = info@projectpanoptes.org
-description = Astronomical utilities. PANOPTES Style.
-edit_on_github = True
-github_project = panoptes/panoptes-utils
-keywords = exoplanet digital dslr camera utilities astronomy
-license = MIT
-long_description = A set of utility modules, functions, and services. Designed for PANOPTES, maybe useful for others.
-name = panoptes-utils
-package_name = panoptes.utils
-url = http://github.com/panoptes/panoptes-utils
-
 [pycodestyle]
 max-line-length = 100
 
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = panoptes/utils/_version.py
-versionfile_build = panoptes/utils/_version.py
-tag_prefix =
-parentdir_prefix = panoptes-utils-
+[options]
+setup_requires =
+    setuptools_scm
